requirements.txt
Django>=5.0
django-allauth>=0.59.0
django-crispy-forms>=2.1
crispy-tailwind>=0.5.0
Pillow>=10.1.0
